% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-wasm.R
\name{is_r_using_wasm}
\alias{is_r_using_wasm}
\title{Detect if version of R is built using Web Assembly (WASM)}
\usage{
is_r_using_wasm()
}
\description{
The function checks whether the compiled version of \emph{R} shows an architecture
that matches \code{"wasm32"}. If its the case, we're likely using webR to access
R.
}
\examples{
# Check to see if WASM is active
is_r_using_wasm()
}
